Added BaseTest Class, splitted tests into Services and Htpp methods
